+++ log-base/src/Log/Backend/StandardOutput/Bulk.hs
@@ -1,34 +1,39 @@
--- | Bulk stdout logging back-end, useful mainly for testing.
-module Log.Backend.StandardOutput.Bulk (
-    withBulkStdOutLogger,
-    bulkStdoutLogger
+-- | Bulk stdout logging back-end.
+module Log.Backend.StandardOutput.Bulk
+  ( withBulkStdOutLogger
+  , withBulkJsonStdOutLogger
   ) where
 
+import Data.Aeson
 import Prelude
-import qualified Data.Text.IO as T
 import System.IO (hFlush, stdout)
+import qualified Data.Text.IO as T
+import qualified Data.ByteString.Lazy.Char8 as BSL
 
 import Log.Data
 import Log.Logger
 import Log.Internal.Logger
 
--- | Create a 'bulkStdoutLogger' for the duration of the given action,
--- and shut it down afterwards, making sure that all buffered messages
--- are actually written to stdout. Flushes 'stdout' on each bulk write.
+-- | Create an asynchronouis logger thread that prints messages to standard
+-- output once per second for the duration of the given action. Flushes 'stdout'
+-- on each bulk write.
 withBulkStdOutLogger :: (Logger -> IO r) -> IO r
 withBulkStdOutLogger act = do
-  logger <- bulkStdoutLogger
+  logger <- mkBulkLogger "stdout-bulk"
+    (\msgs -> do
+        mapM_ (T.putStrLn . showLogMessage Nothing) msgs
+        hFlush stdout
+    ) (return ())
   withLogger logger act
 
-{-# DEPRECATED bulkStdoutLogger "Use 'withBulkStdOutLogger' instead!" #-}
-
--- | Start an asynchronous logger thread that prints messages to
--- standard output.
---
--- Please use 'withBulkStdOutLogger'' instead, which is more exception-safe
--- (see the note attached to 'mkBulkLogger').
-bulkStdoutLogger :: IO Logger
-bulkStdoutLogger = mkBulkLogger "stdout-bulk"
-                   (\msgs -> mapM_ (T.putStrLn . showLogMessage Nothing) msgs
-                             >> hFlush stdout)
-                   (hFlush stdout)
+-- | Create a bulk logger that prints messages in the JSON format to standard
+-- output once per second for the duration of the given action. Flushes 'stdout'
+-- on each bulk write.
+withBulkJsonStdOutLogger :: (Logger -> IO r) -> IO r
+withBulkJsonStdOutLogger act = do
+  logger <- mkBulkLogger "stdout-bulk-json"
+    (\msgs -> do
+        mapM_ (BSL.putStrLn . encode) msgs
+        hFlush stdout
+    ) (return ())
+  withLogger logger act