+++ log-base/CHANGELOG.md
@@ -1,3 +1,11 @@
+# log-base-0.9.0.0 (2020-??-??)
+* Always make data attached to a log message a json object
+* Add unliftio-core-0.2 compatiblity
+* Tidy up flushing stdout in stdout loggers
+* Use simpleStdoutLogger in withSimpleStdOutLogger instead of stdoutLogger
+* Remove deprecated functions
+* Add JSON loggers
+
 # log-base-0.8.0.1 (2020-05-08)
 * Update version bounds.